Rename convert_spec_h5 to spectoh5
