+++ silx/io/spectoh5.py
